fix(iast): invalid line number on insecure cookie report
