+++ ddtrace/appsec/iast/processor.py
@@ -1,4 +1,3 @@
-import json
 from typing import TYPE_CHECKING
 
 import attr
@@ -7,6 +6,7 @@ from ddtrace.appsec._constants import APPSEC
 from ddtrace.appsec._constants import IAST
 from ddtrace.appsec.iast import oce
 from ddtrace.appsec.iast._metrics import _set_metric_iast_request_tainted
+from ddtrace.appsec.iast._utils import _iast_report_to_str
 from ddtrace.appsec.iast._utils import _is_iast_enabled
 from ddtrace.appsec.trace_utils import _asm_manual_keep
 from ddtrace.constants import ORIGIN_KEY
@@ -56,19 +56,9 @@ class AppSecIastSpanProcessor(SpanProcessor):
         data = core.get_item(IAST.CONTEXT_KEY, span=span)
 
         if data:
-            from ddtrace.appsec.iast._taint_tracking import OriginType  # noqa: F401
-            from ddtrace.appsec.iast._taint_tracking._native.taint_tracking import origin_to_str  # noqa: F401
-
-            class OriginTypeEncoder(json.JSONEncoder):
-                def default(self, obj):
-                    if isinstance(obj, OriginType):
-                        # if the obj is uuid, we simply return the value of uuid
-                        return origin_to_str(obj)
-                    return json.JSONEncoder.default(self, obj)
-
             span.set_tag_str(
                 IAST.JSON,
-                json.dumps(attr.asdict(data, filter=lambda attr, x: x is not None), cls=OriginTypeEncoder),
+                _iast_report_to_str(data),
             )
             _asm_manual_keep(span)
