+++ tests/appsec/iast/taint_sinks/test_insecure_cookie.py
@@ -1,4 +1,5 @@
 from ddtrace.appsec._constants import IAST
+from ddtrace.appsec.iast._utils import _iast_report_to_str
 from ddtrace.appsec.iast.constants import VULN_INSECURE_COOKIE
 from ddtrace.appsec.iast.constants import VULN_NO_HTTPONLY_COOKIE
 from ddtrace.appsec.iast.constants import VULN_NO_SAMESITE_COOKIE
@@ -20,6 +21,12 @@ def test_nohttponly_cookies(iast_span_defaults):
     span_report = core.get_item(IAST.CONTEXT_KEY, span=iast_span_defaults)
     assert list(span_report.vulnerabilities)[0].type == VULN_NO_HTTPONLY_COOKIE
     assert list(span_report.vulnerabilities)[0].evidence.value == "foo=bar;secure"
+    assert list(span_report.vulnerabilities)[0].location.line is None
+    assert list(span_report.vulnerabilities)[0].location.path is None
+    str_report = _iast_report_to_str(span_report)
+    # Double check to verify we're not sending an empty key
+    assert '"line"' not in str_report
+    assert '"path"' not in str_report
 
 
 def test_nosamesite_cookies_missing(iast_span_defaults):
